Test of modification of d2b.min.js
I prettified d2b.min.js. It had been a single line. I spaced it out and indentend it.
As a test, I tried to change the color of the borders of the arcs to cyan.
This will check whether I found the right place to do this and whether I can propagate the changes through the
github repository to R to make the change.