Added d2b to this repository since I am going to be using them together
Note that the d2b.min.js file is really minimized including renaming
variables with single letters and functions with single letters
Unreadable.
So this is the full thing and is useful for being able to change.
Note that I'll have to figure out what files d2b.min.js contains and
include those at the end of d2b.R